MVP now complete, can get jokes onmy when passing token as auth.
