+++ auth/authenticate.js
@@ -2,11 +2,12 @@ const jwt = require('jsonwebtoken');
 
 const jwtKey =
   process.env.JWT_SECRET ||
-  'shhh, hush now, love';
+  'add a .env file to root of project with the JWT_SECRET variable';
 
 // quickly see what this file exports
 module.exports = {
   authenticate,
+  generateToken
 };
 
 // implementation details
@@ -26,4 +27,15 @@ function authenticate(req, res, next) {
       error: 'No token provided, must be set on the Authorization Header',
     });
   }
+}
+
+function generateToken(user) {
+  const payload = {
+    subject: user.id,
+    username: user.username
+  };
+  const options = {
+    expiresIn: '1d'
+  }
+  return jwt.sign(payload, jwtKey, options)
 }
